requirements.txt
streamlit
openai==0.28
openai-whisper
pydub
Levenshtein